int>double

+++ debugging-java/src/CalculadoraDeMedidas.java
@@ -19,7 +19,7 @@ public class CalculadoraDeMedidas {
         double soma = 0;
         for(String aluno : alunos){
             System.out.printf("Nota do aluno %s ", aluno);
-            double nota = scanner.nextInt();
+            double nota = scanner.nextDouble();
             soma += nota;
         }
